fix(nd): hide map during mode/range change
