+++ src/instruments/src/ND/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { DisplayUnit } from '@instruments/common/displayUnit';
 import { FlightPlanProvider } from '@instruments/common/flightplan';
 import { useSimVar } from '@instruments/common/simVars';
@@ -64,6 +64,46 @@ const NavigationDisplay: React.FC = () => {
 
     const [opacity] = useSimVar('L:A32NX_MFD_MASK_OPACITY', 'number', 100);
 
+    const [modeChangeShown, setModeChangeShown] = useState(false);
+    const [rangeChangeShown, setRangeChangeShown] = useState(false);
+
+    const firstModeUpdate = useRef(true);
+    const firstRangeUpdate = useRef(true);
+
+    useEffect(() => {
+        if (firstModeUpdate.current) {
+            firstModeUpdate.current = false;
+            return () => {};
+        }
+
+        setModeChangeShown(true);
+
+        const timeout = setTimeout(() => {
+            setModeChangeShown(false);
+        }, 2_500);
+
+        return () => clearTimeout(timeout);
+    }, [modeIndex]);
+
+    useEffect(() => {
+        if (firstRangeUpdate.current) {
+            firstRangeUpdate.current = false;
+            return () => {};
+        }
+
+        // RANGE CHANGE has priority over MODE CHANGE
+        if (modeChangeShown) {
+            setModeChangeShown(false);
+        }
+        setRangeChangeShown(true);
+
+        const timeout = setTimeout(() => {
+            setRangeChangeShown(false);
+        }, 2_500);
+
+        return () => clearTimeout(timeout);
+    }, [rangeIndex]);
+
     return (
         <DisplayUnit
             electricitySimvar={displayIndex === 1 ? 'L:A32NX_ELEC_AC_ESS_BUS_IS_POWERED' : 'L:A32NX_ELEC_AC_2_BUS_IS_POWERED'}
@@ -77,19 +117,39 @@ const NavigationDisplay: React.FC = () => {
                     <WindIndicator adrs={airDataReferenceSource} irs={inertialReferenceSource} />
 
                     {modeIndex === Mode.PLAN && (
-                        <PlanMode rangeSetting={rangeSettings[rangeIndex]} ppos={ppos} efisOption={efisOption} />
+                        <PlanMode
+                            rangeSetting={rangeSettings[rangeIndex]}
+                            ppos={ppos}
+                            efisOption={efisOption}
+                            mapHidden={modeChangeShown || rangeChangeShown}
+                        />
                     )}
                     {modeIndex === Mode.ARC && (
-                        <ArcMode adirsAlign={adirsAlign} rangeSetting={rangeSettings[rangeIndex]} side={side} ppos={ppos} efisOption={efisOption} />
+                        <ArcMode
+                            adirsAlign={adirsAlign}
+                            rangeSetting={rangeSettings[rangeIndex]}
+                            side={side}
+                            ppos={ppos}
+                            efisOption={efisOption}
+                            mapHidden={modeChangeShown || rangeChangeShown}
+                        />
                     )}
                     {(modeIndex === Mode.ROSE_ILS || modeIndex === Mode.ROSE_VOR || modeIndex === Mode.ROSE_NAV)
                     && (
-                        <RoseMode adirsAlign={adirsAlign} rangeSetting={rangeSettings[rangeIndex]} side={side} ppos={ppos} mode={modeIndex} efisOption={efisOption} />
+                        <RoseMode
+                            adirsAlign={adirsAlign}
+                            rangeSetting={rangeSettings[rangeIndex]}
+                            side={side}
+                            ppos={ppos}
+                            mode={modeIndex}
+                            efisOption={efisOption}
+                            mapHidden={modeChangeShown || rangeChangeShown}
+                        />
                     )}
 
                     <Chrono side={side} />
 
-                    <NavigationDisplayMessages adirsAlign={adirsAlign} rangeSetting={rangeSettings[rangeIndex]} mode={modeIndex} />
+                    <NavigationDisplayMessages adirsAlign={adirsAlign} mode={modeIndex} modeChangeShown={modeChangeShown} rangeChangeShown={rangeChangeShown} />
                     {(adirsAlign) && (
                         <>
                             <RadioNavInfo index={1} side={side} />
